home page js
created a pageContent div that contains the
restaurant name (h1), front page image (img), and
restaurant description (p) and append it to the
content div in the html for the home tab. export
it to pageLoad.js where initial load of the page
is conducted and the createRestaurantHomePage
function is called.